'drunk_times': 1000,     'drunk_this_week': 2000,     'drunk_all_time': 4000,     'favorite_drink': 'beer'

+++ db/src/main/java/org/startup/db/UserDao.java
@@ -56,25 +56,25 @@ public class UserDao {
                 new AlcoCoLikerMapper());
     }
 
-    public int getDrunkTimes(long userID) {
+    public Integer getDrunkTimes(long userID) {
         return jdbcTemplate.queryForObject("select count(*) as drunkTimes\n" +
                         "  from alcoholic al JOIN act a ON (al.id = a.alcoholic)\n" +
                         "  where al.id = ?", new Object[]{userID}, Integer.class);
     }
 
-    public long getDrunkThisWeek(long userID) {
+    public Integer getDrunkThisWeek(long userID) {
         return jdbcTemplate.queryForObject("select sum(a.volume)\n" +
                 "  from alcoholic al JOIN act a ON (al.id = a.alcoholic)\n" +
                 "  where al.id = ? and (7*24*60*60 > EXTRACT(EPOCH FROM (current_timestamp - a.ts)))", new Object[]{userID}, Integer.class);
     }
 
-    public long getDrunkAllTime(long userID) {
+    public Integer getDrunkAllTime(long userID) {
         return jdbcTemplate.queryForObject("select sum(a.volume)\n" +
                 "  from alcoholic al JOIN act a ON (al.id = a.alcoholic)\n" +
                 "  where al.id = ?", new Object[]{userID}, Integer.class);
     }
 
-    public long getFavouriteDrink(long userID) {
+    public Integer getFavouriteDrink(long userID) {
         return jdbcTemplate.queryForObject("select a.item, ai.name, count(*) as itemCount\n" +
                 "  from alcoholic al JOIN act a ON (al.id = a.alcoholic)\n" +
                 "    join alcohol_item ai ON (a.item = ai.id)\n" +
